Separa ControlFlowGraph em diferentes arquivos
