+++ src/Compiler/CodeAnalysis/Binding/Binder.cs
@@ -5,6 +5,7 @@ using System.Data;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Text;
+using Compiler.CodeAnalysis.Binding.FlowControl;
 using Compiler.CodeAnalysis.Diagnostics;
 using Compiler.CodeAnalysis.Lowering;
 using Compiler.CodeAnalysis.Symbols;
